requirements.txt
# ============================================
# Sentinel Vision — Dependencies
# ============================================

# Core CV & Vision
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0

# YOLOv8 (Person Detection)
ultralytics>=8.0.0
torch>=2.0.0
torchvision>=0.15.0

# Face Recognition Pipeline
facenet-pytorch>=2.5.3

# Pose Estimation
mediapipe>=0.10.0

# ML / Training
scikit-learn>=1.3.0
joblib>=1.3.0

# GUI
PyQt5>=5.15.0

# Config
pyyaml>=6.0

# Visualization & Analysis
matplotlib>=3.7.0
seaborn>=0.12.0
pandas>=2.0.0

# Notebook
jupyter>=1.0.0
